Create README.md
Exploratory Data Analysis (EDA) on the Titanic Dataset

The Titanic dataset is a well-known dataset used in machine learning and data science for binary classification. It contains information about the passengers aboard the Titanic, which sank in April 1912. The goal of the EDA is to explore the dataset, understand the underlying structure, and extract meaningful insights that could help in predicting whether a passenger survived or not.